+++ src/main/java/lotto/controller/LottoGameController.java
@@ -1,8 +1,8 @@
 package lotto.controller;
 
 import lotto.domain.*;
-import lotto.domain.strategy.ManualGenerateStrategy;
-import lotto.domain.strategy.RandomGenerateStrategy;
+import lotto.strategy.ManualGenerateStrategy;
+import lotto.strategy.RandomGenerateStrategy;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
@@ -15,16 +15,14 @@ public class LottoGameController {
 
     private final InputView inputView;
     private final OutputView outputView;
-    private final LottoTicketMachine ticketMachine;
 
-    public LottoGameController(InputView inputView, OutputView outputView, LottoTicketMachine ticketMachine) {
+    public LottoGameController(InputView inputView, OutputView outputView) {
         this.inputView = inputView;
         this.outputView = outputView;
-        this.ticketMachine = ticketMachine;
     }
 
     public LottoGameController() {
-        this(new InputView(), new OutputView(), new LottoTicketMachine());
+        this(new InputView(), new OutputView());
     }
 
     public void play() {
@@ -48,7 +46,7 @@ public class LottoGameController {
     private WinningTicket getWinningTicket() {
         Set<LottoNumber> winningNumbers = inputView.readLottoNumbers();
         LottoNumber bonusNumber = inputView.readBonusNumber();
-        return ticketMachine.generateWinningTicket(winningNumbers, bonusNumber);
+        return new WinningTicket(winningNumbers, bonusNumber);
     }
 
     private List<Rank> getRanks(List<LottoTicket> lottoTickets, WinningTicket winningTicket) {
@@ -81,11 +79,13 @@ public class LottoGameController {
     }
 
     private List<LottoTicket> getRandomLottoTickets(Money randomTicketPrice) {
-        return ticketMachine.buyLottoTickets(randomTicketPrice, new RandomGenerateStrategy());
+        LottoTicketGenerator randomTicketMachine = new LottoTicketGenerator(new RandomGenerateStrategy());
+        return randomTicketMachine.buyLottoTickets(randomTicketPrice);
     }
 
     private List<LottoTicket> getManualLottoTickets(Money manualTicketPrice) {
-        return ticketMachine.buyLottoTickets(manualTicketPrice, new ManualGenerateStrategy());
+        LottoTicketGenerator manualTicketMahchine = new LottoTicketGenerator(new ManualGenerateStrategy());
+        return manualTicketMahchine.buyLottoTickets(manualTicketPrice);
     }
 
     private Money getMoney() {
