+++ src/main/java/lotto/strategy/RandomGenerateStrategy.java
@@ -1,6 +1,7 @@
-package lotto.domain.strategy;
+package lotto.strategy;
 
 import lotto.domain.LottoNumber;
+import lotto.domain.strategy.LottoNumberGenerateStrategy;
 import lotto.util.NumberGenerateUtil;
 
 import java.util.Set;
