+++ src/test/java/lotto/domain/LottoTicketGeneratorTest.java
@@ -2,11 +2,13 @@ package lotto.domain;
 
 import lotto.exception.InvalidMoneyInputException;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.List;
+import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -36,4 +38,21 @@ class LottoTicketGeneratorTest {
                 .isInstanceOf(InvalidMoneyInputException.class);
     }
 
+    @Test
+    @DisplayName("전략에 따른 로또 티켓 구매")
+    void buyLottoTicketsByStrategyTest() {
+        // given
+        LottoTicketGenerator ticketMachine = new LottoTicketGenerator(new TestGeneratorStrategy());
+        // when
+        List<LottoTicket> lottoTickets = ticketMachine.buyLottoTickets(new Money(1000));
+        // then
+        Set<LottoNumber> lottoNumbers = lottoTickets.get(0).getLottoNumbers();
+        assertThat(lottoNumbers).containsExactly(
+                new LottoNumber(1),
+                new LottoNumber(2),
+                new LottoNumber(3),
+                new LottoNumber(4),
+                new LottoNumber(5),
+                new LottoNumber(6));
+    }
 }
