+++ src/test/java/lotto/domain/RandomGenerateStrategyTest.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.domain.strategy.RandomGenerateStrategy;
+import lotto.strategy.RandomGenerateStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
